backend/requirements.txt
cryptography==3.4.7
Flask==3.0.3
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
pydantic==1.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
Werkzeug==3.0.3
